feat: add survey conversion summary, log, and validation results modules
- Implemented `createSurveyConversionSummaryController` for displaying conversion summaries, including task selection and template matches.
- Added `createSurveyConversionLogController` to manage logging of conversion processes with toggle functionality.
- Introduced `createSurveyValidationResultsController` for displaying validation results, including error and warning summaries with detailed messages.
- Enhanced `survey-value-offset-editor.js` and `survey-workflow-template-check.js` to integrate new functionalities and improve user feedback during conversion processes.
- Updated tests in `test_converter_workflow_wiring.py` to cover new modules and ensure proper wiring of the conversion workflow.
